SEE COMMIT FOR LATEST NOTES!
The problem is:

The massive table cell is receiving a new set of cells because the
hashes of them have been updated.

Why do we send cells to the other cells? For auto-complete. In
reality, we only care about the UUIDs and the names of cells. In
practice, the value of the cells should not matter at all.

So I've changed the equality on output cell to be on name and uuid
alone.

It's still kind of sucky that we re-render an entire cell for the
purpose of at-some-point doing an auto-complete in a sub-editor. I'll
have to have a think about that and what can be done to avoid it.

Another option is just making cell rendering cheaper with incomplete
views and more pagination.